Orders with dishes finally working

+++ PRACTICAS/Practica5/Notepad/app/src/main/java/es/unizar/eina/M42_comidas/database/PedidoPlatoRepository.java
@@ -198,8 +198,8 @@ public class PedidoPlatoRepository {
      * @param esPedido
      * @return un valor entero con el numero de filas modificadas.
      */
-    public int update(EsPedido esPedido) {
-        final int[] result = {0};
+    public long update(EsPedido esPedido) {
+        final long[] result = {0};
         PedidoPlatoRoomDatabase.databaseWriteExecutor.execute(() -> {
             result[0] = mEsPedidoDao.update(esPedido);
         });
